Add files via upload
Week 5 progress was starting the world in the Webots software and get to understand the environment of the software. We were capable of creating part of the code. Majority of the time was to learn how to integrate a new controller into the robot as the issue with the code not working was due to not creating a different controller else than the one that comes with the world. A code to navigate back to the brightest dead-end must be made. The main issue for next week is to be able to figure why the robot is not moving - conditions need adjustments.