Add files via upload
Week 6  - submission of the video progress is this week, and the robot must start moving.  Distance sensors arrays do not show on the robot meaning they are not working. This issue will need to be resolved in week 7. The link to the video progress is: https://youtu.be/Rv_IrPHwWCg   - there is subtitles to the video if needed.