Add files via upload
Week 7 progress had solved the issue. The issue was that we were not compiling the code using the compiler button, which looks like a settings button. Furthermore, the sensors were not added to the ground sensors folder; hence, they were not working. The robot was moving, only rotating in his spot with the improved code for this week - indicating that the conditions still do not allow the performance to be perfected. Furthermore, the function to go to the dead-end with the brightest light was made and added to the code; it can only be tested once the movement conditions are fixed.