Add files via upload
In week 8, we analysed all the possible cases with the sensors of the robot at any intersection type of the maze. According to these cases, we coded them into the program to be able to command the robot what to do, depending on which case. The pdf shows the sketches of the cases explored. The code file shows the controller code programmed for the cases to work. However, the robot confuses many cases due to the maze corridors' thickness.