Add files via upload
In week 9, we explored two methods to fix the issue with the robot. We tried to integrate GPS into the code, which was done successfully. However, the use of the GPS was only to obtain coordinates of the dead-end with the brightest light intensity and go there. However, in the backtracking condition, the robot seemed not to move at all as it could not backtrack by following the maze again. We tried using switch cases to refer to them easier which did help the robot with navigating around the maze in the first cycle. The backtracking stage is an issue that needs to be resolved in Week 10.