Add files via upload
After researching maze-solving algorithms and asking other groups the logic they found best, we decided to integrate the wall-following algorithm due to the robot getting confused with the scenario it must execute. For the backtracking issue, a second ground sensor was added to the robot to be able to store values for the second maze cycle. Then, two different functions were added to the code - one to follow the maze walls with the backtracking conditions  - when the dead-end with the brightest light is found, the robot must stop, if not, it continues following the wall. Moreover, we included a bool function that measures the light at the dead-end and compares the value of the light obtained with the known maximum light intensity found in the first maze cycle - returns true if the light obtained is greater or equal to the known. We had to use tolerance value to exclude the noise as the robot will not move in the backtracking state due to the values in the second maze lap for the intensities being slightly greater.